src/prompts/prompt_templates.py
from langchain_core.prompts import ChatPromptTemplate

def get_system_prompt():
    """
    Returns the system prompt template.
    """
    template = (
        "You are an assistant for question-answering tasks. "
        "Use the following pieces of retrieved context to answer "
        "the question. If you don't know the answer, say that you "
        "don't know. Use three sentences maximum and keep the "
        "answer concise."
        "\n\n"
        "{context}"
    )
    return ("system", template)

def get_user_prompt():
    """
    Returns the user prompt template.
    """
    return ("human", "{input}")
