requirements.txt
langchain
langchain-openai
langchain-classic
langchain-community
pypdf
faiss-cpu
sentence-transformers
python-dotenv
